Trim example-utils
