+++ examples/utils/example-utils/src/index.ts
@@ -21,13 +21,7 @@ export {
 } from "./containerViewRuntimeFactory.js";
 export type {
 	DataTransformationCallback,
-	IAcceptedMigrationDetails,
 	IImportExportModel,
-	IMigratableModel,
-	IMigrationTool,
-	IMigrationToolEvents,
-	IMigrator,
-	IMigratorEvents,
 	ISameContainerMigratableModel,
 	ISameContainerMigratableModelEvents,
 	ISameContainerMigrationTool,
@@ -35,27 +29,17 @@ export type {
 	ISameContainerMigrator,
 	ISameContainerMigratorEvents,
 	IVersionedModel,
-	MigrationState,
 	SameContainerMigrationState,
 } from "./migrationInterfaces/index.js";
 export {
-	MigrationToolFactory,
 	SameContainerMigrationTool,
 	SameContainerMigrationToolInstantiationFactory,
 } from "./migrationTool/index.js";
-export { Migrator, SameContainerMigrator } from "./migrator/index.js";
+export { SameContainerMigrator } from "./migrator/index.js";
 export {
-	CreateModelCallback,
-	IAttachedMigratableModel,
-	IDetachedMigratableModel,
 	IDetachedModel,
-	IMigratableModelContainerRuntimeEntryPoint,
-	IMigratableModelLoader,
 	IModelContainerRuntimeEntryPoint,
 	IModelLoader,
-	instantiateMigratableRuntime,
-	MigratableModelLoader,
-	MigratableSessionStorageModelLoader,
 	ModelContainerRuntimeFactory,
 	ModelLoader,
 	SessionStorageModelLoader,
