+++ examples/utils/example-utils/src/migrationInterfaces/index.ts
@@ -5,23 +5,9 @@
 
 export {
 	IImportExportModel,
-	IMigratableModel,
-	IVersionedModel,
-} from "./migratableModel.js";
-export {
-	IAcceptedMigrationDetails,
-	IMigrationTool,
-	IMigrationToolEvents,
-	MigrationState,
-} from "./migrationTool.js";
-export {
-	DataTransformationCallback,
-	IMigrator,
-	IMigratorEvents,
-} from "./migrator.js";
-export {
 	ISameContainerMigratableModel,
 	ISameContainerMigratableModelEvents,
+	IVersionedModel,
 } from "./sameContainerMigratableModel.js";
 export {
 	ISameContainerMigrationTool,
@@ -29,6 +15,7 @@ export {
 	SameContainerMigrationState,
 } from "./sameContainerMigrationTool.js";
 export {
+	DataTransformationCallback,
 	ISameContainerMigrator,
 	ISameContainerMigratorEvents,
 } from "./sameContainerMigrator.js";
